PDF Support, go bindings, and github actions (#4)
* fix: align ExifData.data type to *mut u8 to match C header

The C FFI header declares ExifData.data as `unsigned char *` but the Rust
struct used `*mut u32`. Fix the field type and update from_bytes() and
from_exif_value() to use *mut u8 throughout, removing the incorrect u32
casts and Box::into_raw usage.

* feat: add page_count field to AgnoImage struct and C header

Add `page_count: u64` to `AgnoImage` between `height` and `exif` (required
for #[repr(C)] FFI field order). Default to 1 in both constructors. Add
`load_pdf_page` declaration to `lib/agno.h`.

* feat: initialize Go bindings module structure

* feat: add typed ExifTag constants for Go bindings

* add go bindings

* add workflows

* WIP

* WIP

* WIP

* WIP

+++ agno/src/canon_decoder.rs
@@ -487,8 +487,8 @@ fn decode_ljpeg_data<R: Read>(
     // We need to store the first value of each component from the previous row
     // to use as the predictor for the first column of subsequent rows.
 
-    let mut pred = vec![initial_pred as i32; num_comps];
-    let mut row_start_pred = vec![initial_pred as i32; num_comps]; // First value of each component in prev row
+    let mut pred = vec![initial_pred; num_comps];
+    let mut row_start_pred = vec![initial_pred; num_comps]; // First value of each component in prev row
     let samples_per_slice = cr2_slice_1 * jh_high;
 
     for jrow in 0..jh_high {
@@ -507,7 +507,7 @@ fn decode_ljpeg_data<R: Read>(
             // Check for restart marker (resets decoder state)
             if bs.restart_pending {
                 for p in pred.iter_mut() {
-                    *p = initial_pred as i32;
+                    *p = initial_pred;
                 }
                 bs.reset();
             }
